# static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Maze Bank API",
    "description": "API documentation for the deliberately vulnerable banking application",
    "version": "1.0.0",
    "contact": {
      "name": "Maze Bank Support",
      "url": "https://mazebank.com"
    }
  },
  "servers": [
    {
      "url": "https://mazebank.com",
      "description": "Controlled Production Server"
    }
  ],
...
  "paths": {
...
    "/login": {
      "post": {
        "tags": ["authentication"],
        "summary": "Log in to the application",
        "description": "Authenticate user and receive JWT token. Vulnerable to SQL injection.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "testuser"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/transfer": {
      "post": {
        "tags": ["transactions"],
        "summary": "Transfer money",
        "description": "Transfer money between accounts. Vulnerable to race conditions.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to_account": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string",
                    "required": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfer successful"
          }
        }
      }
    },
...
    "/check_balance/{account_number}": {
      "get": {
        "tags": ["transactions"],
        "summary": "Check account balance",
        "description": "Check balance of any account. Vulnerable to BOLA.",
        "parameters": [
          {
            "name": "account_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Balance retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "balance": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Account not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Vulnerable JWT implementation"
      }
    }
  }
}